# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: TASK EXAMPLE API REST
  description: 'EJEMPLO DE API REST CON SPRINGBOOT '
  contact:
    name: SENISE
  license:
    name: SIN LICENCIA 😂
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- Bearer_Authentication: []
...
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: Obtener detalles de un usuario
      description: Este endpoint devuelve los detalles de un usuario específico por
        su ID.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: ID del usuario a obtener
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Usuario encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: Usuario no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
    put:
      tags:
      - Users
      summary: Actualizar usuario existente
      description: Este endpoint permite actualizar un usuario existente por su ID.
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: ID del usuario que se desea actualizar.
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestDTO"
        required: true
      responses:
        "200":
          description: Usuario actualizado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Datos inválidos proporcionados
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: Usuario no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
    delete:
      tags:
      - Users
      summary: Eliminar un usuario
      description: Este endpoint permite eliminar un usuario por su ID.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: ID del usuario a eliminar
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Usuario eliminado con éxito
        "404":
          description: Usuario no encontrado
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: Obtener todos los usuarios
      description: Este endpoint devuelve una lista con todos los usuarios almacenados
        en el sistema.
      operationId: getAllUsers
      responses:
        "200":
          description: Lista de usuarios obtenida con éxito
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDTO"
    post:
      tags:
      - Users
      summary: Crear un nuevo usuario
      description: Este endpoint permite crear un nuevo usuario en el sistema.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestDTO"
        required: true
      responses:
        "201":
          description: Usuario creado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Datos inválidos proporcionados
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
...
  /api/v1/users/search:
    get:
      tags:
      - Users
      summary: Buscar usuarios por múltiples criterios
      description: "Permite buscar usuarios por nombre, email o rol. Todos los pará\
        metros son opcionales, y se puede combinar cualquiera de ellos."
      operationId: searchUsers
      parameters:
      - name: name
        in: query
        description: Nombre del usuario
        required: false
        schema:
          type: string
      - name: email
        in: query
        description: Email del usuario
        required: false
        schema:
          type: string
      - name: role
        in: query
        description: Rol del usuario
        required: false
        schema:
          type: string
      responses:
        "400":
          description: Parámetros de búsqueda inválidos
        "200":
          description: Búsqueda realizada con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDTO"
        "500":
          description: Error interno del servidor
        "204":
          description: No se encontraron usuarios que coincidan con los criterios
            de búsqueda
  /api/v1/users/created-in-period:
    get:
      tags:
      - Users
      summary: Buscar usuarios creados en un periodo de tiempo
      description: Permite buscar usuarios creados dentro de un rango de fechas. Los
        parámetros de fecha son obligatorios.
      operationId: getUsersCreatedInPeriod
      parameters:
      - name: startDate
        in: query
        description: Fecha de inicio del periodo
        required: true
        schema:
          type: string
          format: date-time
      - name: endDate
        in: query
        description: Fecha de fin del periodo
        required: true
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: Búsqueda realizada con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDTO"
        "500":
          description: Error interno del servidor
        "400":
          description: Parámetros de fecha inválidos
  /api/v1/tasks/user/{userId}:
    get:
      tags:
      - Tasks
      summary: Obtener todas las tareas de un usuario
      description: Este endpoint devuelve todas las tareas asociadas al usuario con
        el ID especificado.
      operationId: getTasksByUserId
      parameters:
      - name: userId
        in: path
        description: ID del usuario cuyas tareas se desean obtener
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponseDTO"
...
components:
  schemas:
    UserRequestDTO:
      required:
      - email
      - name
      - password
      - role
      type: object
      properties:
        name:
          type: string
          description: Nombre del usuario
          example: Juan Pérez
        email:
          type: string
          description: Correo electrónico del usuario
          example: juan.perez@example.com
        password:
          type: string
          description: Contraseña del usuario
          example: Abc1234.
        role:
          type: string
          description: Rol del usuario
          example: ROLE_USER
      description: Detalles del usuario a actualizar
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID del usuario
          format: int64
          example: 1
        name:
          type: string
          description: Nombre del usuario
          example: Juan Pérez
        email:
          type: string
          description: Correo electrónico del usuario
          example: juan.perez@example.com
        role:
          type: string
          description: Rol del usuario
          example: ROLE_USER
        createdAt:
          type: string
          description: Fecha de creación
          format: date-time
      description: DTO para la respuesta de un usuario
...
    TaskResponseDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID único de la tarea
          format: int64
          example: 1
        title:
          type: string
          description: Título de la tarea
          example: Comprar leche
        description:
          type: string
          description: Descripción detallada de la tarea
          example: Ir al supermercado a comprar leche
        completed:
          type: boolean
          description: Indica si la tarea está completada
          example: false
        name:
          type: string
          description: Nombre del usuario asignado a la tarea
          example: Juan Pérez
        createdAt:
          type: string
          description: Fecha de creación
          format: date-time
      description: DTO que representa la tarea en la respuesta
...
  securitySchemes:
    Bearer_Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT